convert most of openbsd-compat to conditional build rather than conditionally ifdef out bits
