+++ openbsd-compat/fgetln.c
@@ -18,7 +18,6 @@
  * portable fgetln() version, NOT reentrant
  */
 
-#ifndef HAVE_FGETLN
 #include <stdio.h>
 #include <stdlib.h>
 #include <errno.h>
@@ -57,4 +56,4 @@ fgetln(FILE *fp, size_t *len)
 	}
 	return (*len = r) ? buf : NULL;
 }
-#endif
+
