+++ openbsd-compat/freezero.c
@@ -21,7 +21,6 @@
 
 #include "includes.h"
 
-#ifndef HAVE_FREEZERO
 #include <stdlib.h>
 
 void
@@ -32,5 +31,3 @@ freezero(void *ptr, size_t sz)
 	explicit_bzero(ptr, sz);
 	free(ptr);
 }
-
-#endif /* HAVE_FREEZERO */
