+++ openbsd-compat/reallocarray.c
@@ -18,7 +18,7 @@
 /* OPENBSD ORIGINAL: lib/libc/stdlib/reallocarray.c */
 
 #include "includes.h"
-#ifndef HAVE_REALLOCARRAY
+
 #include <sys/types.h>
 #include <errno.h>
 #include <stdint.h>
@@ -40,4 +40,3 @@ reallocarray(void *optr, size_t nmemb, size_t size)
 	}
 	return realloc(optr, size * nmemb);
 }
-#endif
