+++ openbsd-compat/strlcpy.c
@@ -19,7 +19,6 @@
 /* OPENBSD ORIGINAL: lib/libc/string/strlcpy.c */
 
 #include "includes.h"
-#ifndef HAVE_STRLCPY
 
 #include <sys/types.h>
 #include <string.h>
@@ -54,5 +53,3 @@ strlcpy(char *dst, const char *src, size_t siz)
 
 	return(s - src - 1);	/* count does not include NUL */
 }
-
-#endif /* !HAVE_STRLCPY */
